Moved command tooltip to runtime folder and the custom editor in it's own script. Furthermore, the tooltip attribute will only work on scriptable objects of type ConsoleCommand
